Added JSON files for shipping container and load them
The one more complex container has messed up images. I am thinking it is an issue in the JSON might just not be able to use it. The other one loads in just fine. I am not able to move them around. Thinking about setting a var in the lambda function so we can reference the object outside of it. Not sure though